+++ lending/RELEASES.md
@@ -569,3 +569,13 @@ Based on:
 - [typescript v0.6.5] lending
 ### Releases
 - [NPM v0.6.5] https://www.npmjs.com/package/ryan-lending/v/0.6.5 - lending
+
+## 2024-11-12 00:57:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.438.1 (2.457.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.6] lending
+### Releases
+- [NPM v0.6.6] https://www.npmjs.com/package/ryan-lending/v/0.6.6 - lending